+++ dev_history.R
@@ -41,3 +41,19 @@ undebug(
     files_1985
   )
 )
+
+
+# 2026-03-21 ====
+
+## DOCUMENT, CHECK, AND BUILD PACKAGE ====
+
+devtools::document()
+devtools::check()
+# devtools::build()
+pak::local_install(ask = FALSE)
+
+tinytest::run_test_dir()
+covr::report()
+
+# remotes::install_git("https://codeberg.org/tim-salabim/oiseasy.git")
+oiseasy::bumpDevVersion()